requirements.txt
fastapi[standard]
python-dotenv
httpx
